Finished Auth and Migrations

+++ internal/models/order.go
@@ -3,19 +3,19 @@ package models
 import "time"
 
 type Order struct {
-    ID        int       `json:"id"`
-    UserID    int       `json:"user_id"`
-    Total     float64   `json:"total"`
-    CreatedAt time.Time `json:"created_at"`
-    UpdatedAt time.Time `json:"updated_at"`
+	ID        int       `json:"id"`
+	UserID    int       `json:"user_id"`
+	Total     float64   `json:"total"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
 }
 
 type OrderItem struct {
-    ID        int       `json:"id"`
-    OrderID   int       `json:"order_id"`
-    ProductID int       `json:"product_id"`
-    Quantity  int       `json:"quantity"`
-    Price     float64   `json:"price"`
-    CreatedAt time.Time `json:"created_at"`
-    UpdatedAt time.Time `json:"updated_at"`
+	ID        int       `json:"id"`
+	OrderID   int       `json:"order_id"`
+	ProductID int64     `json:"product_id"`
+	Quantity  int       `json:"quantity"`
+	Price     float64   `json:"price"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
 }
